+++ src/main/java/fr/dwightstudio/jarmemu/sim/StateContainer.java
@@ -5,7 +5,7 @@ import java.util.HashMap;
 public class StateContainer {
 
     // ASM
-    public final HashMap<String, Byte> symbols; // HashMap des symbols càd des labels et des variables ASM (du .data)
+    public final HashMap<String, Integer> symbols; // HashMap des symbols càd des labels et des variables ASM (du .data)
 
     // Registers
     public static final int REGISTER_NUMBER = 16;
